backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
email-validator
bcrypt
python-jose[cryptography]
python-multipart
pgvector
